makefile: Fixes for building with LLVM
This commit fixes building with LLVM by adding a way to change from GCC to Clang (and other Binutils/LLVM binaries) depending on the environnement variable "LLVM", when set to 1, the LLVM tools are enabled otherwise, Binutils is used to build the kernel.

This commit also fixes code that isn't compatible with Clang.