drvgen: Use python2.7 as default
Back in the days of 4.9 kernel, MediaTek made this DrvGen.py script without future proofing. When Python 3 came out (with a lot of changes), this script was broken and because MediaTek seems to be dumb, they hardcoded the "python" path instead of "python2.7", so this script breaks if you don't have Python 2 currently installed. the Python version might differ depending on your distribution, in this case im building use Debian so it should work as long as your distro uses "python2.7" instead of "python2".